upd. 支持使用腾讯云存储图片数据

+++ functions/deleteFiles/index.js
@@ -2,6 +2,9 @@ module.exports = async (ctx) => {
     const fileIDs = ctx.args.photoIDs
     if (fileIDs && fileIDs.length > 0) {
         for (var idx in fileIDs) {
+            if (!fileIDs[idx].includes('cdn.bspapp.com')) {
+                continue;
+            }
             await ctx.mpserverless.file.deleteFile(fileIDs[idx])
         }
         return { success: true, deletedCount: fileIDs.length };
